+++ modules/files_folders_functions/builtin/cpp/cdBuiltin.cpp
@@ -27,11 +27,12 @@
 #include "ChangeDirectory.hpp"
 #include "Error.hpp"
 #include "GetCurrentDirectory.hpp"
+#include "NelsonPrint.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
 ArrayOfVector
-Nelson::FilesFoldersGateway::cdBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+Nelson::FilesFoldersGateway::cdBuiltin(int nLhs, const ArrayOfVector& argIn)
 {
     ArrayOfVector retval;
     if (argIn.size() > 1) {
@@ -46,10 +47,7 @@ Nelson::FilesFoldersGateway::cdBuiltin(Evaluator* eval, int nLhs, const ArrayOfV
             Error(_W("Impossible to get current directory."));
         } else {
             if (nLhs == 0) {
-                Interface* io = eval->getInterface();
-                if (io) {
-                    io->outputMessage(pwd);
-                }
+                NelsonPrint(pwd);
             } else {
                 retval.push_back(ArrayOf::characterArrayConstructor(pwd));
             }