+++ modules/files_folders_functions/builtin/include/cdBuiltin.hpp
@@ -26,12 +26,11 @@
 #pragma once
 //=============================================================================
 #include "ArrayOf.hpp"
-#include "Evaluator.hpp"
 //=============================================================================
 namespace Nelson {
 namespace FilesFoldersGateway {
     ArrayOfVector
-    cdBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+    cdBuiltin(int nLhs, const ArrayOfVector& argIn);
 }
 } // namespace Nelson
 //=============================================================================