+++ modules/functions_manager/builtin/include/whichBuiltin.hpp
@@ -31,7 +31,7 @@
 namespace Nelson {
 namespace FunctionsGateway {
     ArrayOfVector
-    whichBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+    whichBuiltin(int nLhs, const ArrayOfVector& argIn);
 }
 } // namespace Nelson
 //=============================================================================